Update timelinejs.html
Removed the "{{ with $VARIABLE }}" articles from the file, and instead went with direct variable injection. 

This was due to discovering a ";", a semicolon, was clipped from the last line of the timeline initiation script, this caused the JavaScript to fail when loading. New formatting should produce a more consistent result.